requirements.txt
flask
     opencv-python
     numpy
     pillow